Remove all engines and fetch resource files
