--- engines-example-aarch64.ini
@@ -1,39 +1,0 @@
-[NNUE/StockfishNN]
-name = StockfishNN 17
-web = Stockfish NNUE 17
-small = Stf NN
-medium = Stocf NN
-large = StockfishNN
-elo = 3640
-ccrl = 3640
-levels = 20 + auto, 18 1 core, levels Elo@3000 and Elo@3100 2 cores, level Elo@3190 3 cores
-ponder/brain = y
-fischer chess960 support = y
-comments: Specify NNUE network files in the <engine>.uci file. Uses embedded nets as default.
-
-[a-stockf]
-name = Stockfish 16
-web = Stockfish 16 Classic
-small = Stf 16
-medium = Stockf16
-large = Stockfish16
-elo = 3541
-ccrl = 3541
-levels = 20 + auto, 18 1 core, levels Elo@3000 and Elo@3100 2 cores, level Elo@3190 3 cores
-ponder/brain = y
-fischer chess960 support = y
-comments: No embedded net and NNUE off. Default net (nn-5af11540bbfe.nnue), if needed is in NNUE folder.
-
-[lc0v032]
-name = Lc0 v0.32
-web = Lc0 v0.32
-small = Lc032
-medium = Lc0v032
-large = Lc0 v0.32
-elo = 2400+
-ccrl = 3464 with a larger weights file & GPU
-levels = 4
-cores = 1-4 selectable in level
-ponder/brain = y
-fischer chess960 support = y
-comments: better with 'no book' so it can play learnt openings, bad endgame play
