--- engines-example-x86_64.ini
@@ -1,14 +1,0 @@
-[a-stockf]
-name = Stockfish 17
-small = SF 17
-medium = Stocf 17
-large = Stockfish17
-elo = 3516
-
-[b-lc0]
-name = Leela 032
-small = LC0
-medium = LC0 032
-large = LeelaChessZero 032
-elo = 3516
-
