+++ engines/pgn_engine/pgn_engine.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd /opt/picochess/engines/aarch64/extra
+/opt/picochess/venv/bin/python3 pgn_engine.py
