--- engines/x86_64/.gitignore
@@ -1,3 +1,0 @@
-# Ignore any engines out of the first 8 [a-h] ones
-[i-z]-*
-games.pgn
