--- engines/x86_64/extra/pgn_engine
@@ -1,3 +1,0 @@
-#!/bin/bash
-cd /opt/picochess/engines/x86_64/extra
-/opt/picochess/venv/bin/python3 pgn_engine.py
