--- favorites-example-x86_64.ini
@@ -1,14 +1,0 @@
-; *****************************************
-; Extra Engines
-; *****************************************
-
-[extra/pgn_engine]
-name = PGN Replay & Analysis Engine
-small = PGN
-medium = PGN
-large = PGN Replay
-web = PGN Replay & Analysis
-elo = 0
-levels = 1
-ponder/brain = n
-
