--- pgn_engine-log.txt
@@ -1,81 +1,0 @@
-*** uci
-< id name PGN Replay/Analysis Engine V1.0
-< id author Molli
-< option name pgn_game_file type string default /opt/picochess/games/last_game.pgn
-< option name game_sequence type string default random
-< option name pgn_audio_file type string default 
-< option name max_guess type spin default 0 min 0 max 10
-< option name engine_path type string default /opt/picochess/engines/armv7l/a-stockf
-< option name think_time type spin default 3 min 0 max 60
-< uciok
-*** setoption name pgn_game_file value /opt/picochess/engines/pgn_engine/pgn_games/hoerspiel_nwz.pgn
-< # pgn_game_file: /opt/picochess/engines/pgn_engine/pgn_games/hoerspiel_nwz.pgn
-*** setoption name pgn_audio_file value /opt/picochess/engines/pgn_engine/pgn_audio/hoerspiel_nwz.ogg
-< # pgn_audio_file: /opt/picochess/engines/pgn_engine/pgn_audio/hoerspiel_nwz.ogg
-*** setoption name max_guess value 0
-< # max_guess: 0
-*** setoption name engine_path value /opt/picochess/engines/armv7l/a-stockf
-< # engine_path: /opt/picochess/engines/armv7l/a-stockf
-*** setoption name think_time value 1
-< # think_time: 1
-*** isready
-Game(s) from PGN file loaded.
-max_games 1
-game_counter 1
-< id name PGN Replay/Analysis Engine V1.0
-< readyok
-*** ucinewgame
-game index: 0
-FEN: 
-Next game no. from PGN: 0
-Number of moves: 140
-*** isready
-Game(s) from PGN file loaded.
-max_games 1
-game_counter 1
-< id name PGN Replay/Analysis Engine V1.0
-< readyok
-*** position startpos
-position startpos ready
-*** setoption name Ponder value false
-*** setoption name UCI_AnalyseMode value false
-*** isready
-Game(s) from PGN file loaded.
-max_games 1
-game_counter 1
-< id name PGN Replay/Analysis Engine V1.0
-< readyok
-*** uci
-pause audio
-< id name PGN Replay/Analysis Engine V1.0
-< id author Molli
-< option name pgn_game_file type string default /opt/picochess/games/last_game.pgn
-< option name game_sequence type string default random
-< option name pgn_audio_file type string default 
-< option name max_guess type spin default 0 min 0 max 10
-< option name engine_path type string default /opt/picochess/engines/armv7l/a-stockf
-< option name think_time type spin default 3 min 0 max 60
-< uciok
-*** uci
-continue audio
-< id name PGN Replay/Analysis Engine V1.0
-< id author Molli
-< option name pgn_game_file type string default /opt/picochess/games/last_game.pgn
-< option name game_sequence type string default random
-< option name pgn_audio_file type string default 
-< option name max_guess type spin default 0 min 0 max 10
-< option name engine_path type string default /opt/picochess/engines/armv7l/a-stockf
-< option name think_time type spin default 3 min 0 max 60
-< uciok
-*** uci
-pause audio
-< id name PGN Replay/Analysis Engine V1.0
-< id author Molli
-< option name pgn_game_file type string default /opt/picochess/games/last_game.pgn
-< option name game_sequence type string default random
-< option name pgn_audio_file type string default 
-< option name max_guess type spin default 0 min 0 max 10
-< option name engine_path type string default /opt/picochess/engines/armv7l/a-stockf
-< option name think_time type spin default 3 min 0 max 60
-< uciok
-*** quit
