--- pgn_game_info.txt
@@ -1,7 +1,0 @@
-PGN_GAME=Kasparov 25 board Simul
-PGN_GAME_INDEX=1
-PGN_PROBLEM=
-PGN_White=Kasparov
-PGN_Black=Lloyds Bank: Munro, Clarke, Cooper, Drew
-PGN_FEN=
-PGN_RESULT=1/2-1/2
