+++ utilities.py
@@ -40,7 +40,7 @@ from typing import Optional
 from pathlib import Path
 
 # picochess version
-version = "4.1.4"
+version = "4.1.5"
 
 logger = logging.getLogger(__name__)
