+++ web/picoweb/templates/clock.html
@@ -116,7 +116,7 @@
                             <div class="card">
                                 <div class="card-header"
                                     style="font-size: 2.5vh; display: flex; justify-content: space-between; align-items: center;">
-                                    <span>PicoChess Version 4.1.4</span>
+                                    <span>PicoChess Version 4.1.5</span>
                                     <div style="display: flex; align-items: center; gap: 8px;">
                                         <form id="moveForm" autocomplete="off" style="margin: 0;">
                                             <input id="moveInput" type="text" inputmode="none" style="
